Changed shard id type to int64 (#296)
With namespaces, we'll be able to create and delete shards. It will be
safer to use `int64` rather than `uint32` to avoid having to deal with
any rollover.

+++ common/shards_test.go
@@ -22,7 +22,7 @@ import (
 
 func TestGenerateShards(t *testing.T) {
 	type args struct {
-		baseId    uint32
+		baseId    int64
 		numShards uint32
 	}
 	tests := []struct {
